api/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
mysql-connector-python==8.1.0
pylint==3.0.2
Flask-Pydantic==0.11.0